+++ NEWS.md
@@ -6,6 +6,7 @@
  * The history sidebar now works again
  * TextCells with copy-alt-text caused an crash
  * A few additional small bugfixes
+ * in sqrt(conjugate(f(x))) the horizontal lines coincided
  * Strings were searched for comment starts when adding line endings
  * Faster start-up by avoiding unneeded redraw events
  * A big number of performance improvements